pyproject.toml
[project]
name = "gambly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hikari-arc>=2.1.1",
    "hikari-miru>=4.2.0",
    "pymongo>=4.14.0",
    "python-dotenv>=1.1.1",
]
